Update pipeline and trigger manifests
This update removes the embedded resource specs within the trigger
template resource and instead uses resource references to 3 pipeline
resources. 2 of those are defined separately in pipeline.yaml since they
are static. The 3rd is added to the resourceTemplates field within
the trigger template resource that is parameterized based on the bindings to the
trigger template.

This is necessary because there's currently a bug that prevents
embedded resource specs within the trigger template from successfully
being copied over to the pipelinerun that is auto-generated when you
manually rerun the pipeline from the OpenShift Console.